move enums to separate file to prevent circular imports
