+++ music_assistant/controllers/music/artists.py
@@ -4,7 +4,6 @@ import asyncio
 import itertools
 from typing import List
 
-from music_assistant.constants import EventType, MassEvent
 from music_assistant.helpers.compare import (
     compare_album,
     compare_strings,
@@ -13,6 +12,8 @@ from music_assistant.helpers.compare import (
 from music_assistant.helpers.database import TABLE_ARTISTS
 from music_assistant.helpers.json import json_serializer
 from music_assistant.helpers.util import create_sort_name
+from music_assistant.models.enums import EventType
+from music_assistant.models.event import MassEvent
 from music_assistant.models.media_controller import MediaControllerBase
 from music_assistant.models.media_items import (
     Album,
