+++ music_assistant/controllers/music/playlists.py
@@ -4,13 +4,14 @@ from __future__ import annotations
 from time import time
 from typing import List
 
-from music_assistant.constants import EventType, MassEvent
 from music_assistant.helpers.database import TABLE_PLAYLISTS
 from music_assistant.helpers.json import json_serializer
 from music_assistant.helpers.util import create_sort_name
+from music_assistant.models.enums import EventType, MediaType
 from music_assistant.models.errors import InvalidDataError, MediaNotFoundError
+from music_assistant.models.event import MassEvent
 from music_assistant.models.media_controller import MediaControllerBase
-from music_assistant.models.media_items import MediaType, Playlist, Track
+from music_assistant.models.media_items import Playlist, Track
 
 
 class PlaylistController(MediaControllerBase[Playlist]):
