+++ music_assistant/controllers/music/tracks.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 import asyncio
 from typing import List
 
-from music_assistant.constants import EventType, MassEvent
 from music_assistant.helpers.compare import (
     compare_artists,
     compare_strings,
@@ -13,8 +12,10 @@ from music_assistant.helpers.compare import (
 from music_assistant.helpers.database import TABLE_TRACKS
 from music_assistant.helpers.json import json_serializer
 from music_assistant.helpers.util import create_sort_name
+from music_assistant.models.enums import EventType, MediaType
+from music_assistant.models.event import MassEvent
 from music_assistant.models.media_controller import MediaControllerBase
-from music_assistant.models.media_items import ItemMapping, MediaType, Track
+from music_assistant.models.media_items import ItemMapping, Track
 
 
 class TracksController(MediaControllerBase[Track]):
