+++ music_assistant/providers/qobuz.py
@@ -10,13 +10,14 @@ from typing import List, Optional
 import aiohttp
 from asyncio_throttle import Throttler
 
-from music_assistant.constants import EventType, MassEvent
 from music_assistant.helpers.app_vars import (  # pylint: disable=no-name-in-module
     app_var,
 )
 from music_assistant.helpers.cache import use_cache
 from music_assistant.helpers.util import parse_title_and_version, try_parse_int
+from music_assistant.models.enums import EventType
 from music_assistant.models.errors import LoginFailed
+from music_assistant.models.event import MassEvent
 from music_assistant.models.media_items import (
     Album,
     AlbumType,
